yBooking-Backend/requirements.txt
Flask
Flask-Cors
pymysql
gunicorn
python-dotenv